refactor: split app into routes and layout
