+++ src/pages/AIChatPage.tsx
@@ -1,0 +1,5 @@
+import AIChat from '../components/AIChat';
+
+const AIChatPage = () => <AIChat />;
+
+export default AIChatPage;
